backend/requirements.txt
django
django-cors-headers
djangorestframework
django-allauth
djangorestframework-simplejwt
celery[redis]
django-celery-beat
mysqlclient
spotipy
cryptography